Fix up old tests and add new ones.

+++ app/models/meta_tagging.rb
@@ -32,7 +32,7 @@ class MetaTagging < ApplicationRecord
   end
 
   def update_inherited
-    sub_tag.async(:update_inherited_meta_tags) if direct
+    sub_tag.async(:update_inherited_meta_tags) if direct && sub_tag
   end
 
   def expire_caching
@@ -47,18 +47,7 @@ class MetaTagging < ApplicationRecord
       # Let callers do something on each iteration.
       yield mt, valid if block_given?
 
-      next if valid
-
-      if mt.sub_tag && mt.meta_tag
-        # We use this method instead of mt.destroy because we want to trigger the
-        # before_remove callbacks on mt.sub_tag, thus ensuring that we clean up
-        # the filter_taggings associated with this MetaTagging.
-        mt.sub_tag.meta_tags.delete(mt.meta_tag)
-      else
-        # But in this case, one of the two tags is missing, so we can only
-        # properly delete the meta tagging by calling mt.destroy:
-        mt.destroy
-      end
+      mt.destroy unless valid
     end
   end
 end
